+++ Eshop_ProjetSpring/src/main/java/fr/adaming/dao/IProduitDao.java
@@ -6,9 +6,9 @@ import fr.adaming.model.Produit;
 
 public interface IProduitDao extends IGeneriqueDao<Produit>{
 	
-	public List<Produit> consulterParCategorie(int idCategorie);
+	public List<Produit> consulterParCategorie(Long idCategorie);
 	
-	public List<Produit> consulterSelection(int idClient);
+	public List<Produit> consulterSelection(Long idClient);
 	
 	public List<Produit> consulterParMotClef(String designation);
